Task A5 (инкремент 4): устранение CS-предупреждений + zero-NEW-warning baseline (66 to 0)
CS0414/0675/8618/8765/8600-8625 исправлены; Puncher OpenWriter refactor; baseline 0 CS (допущено только уменьшение). build 0 CS, check PASS exit 0, fast 3/3 OK, Python 40 OK.

+++ src/besm6.net/Core/Word48Extensions.cs
@@ -7,7 +7,7 @@ public static class Word48Extensions
         ulong val = 0;
         for (int i = 0; i < oct.Length; i++)
         {
-            val = (val << 3) | (ulong)(oct[i] - '0');
+            val = (val << 3) | (uint)(oct[i] - '0');
         }
         return new Word48(val);
     }
